requirements.txt
fastapi
uvicorn
DrissionPage
pymysql
pydantic
python-dotenv
